flipped the neuroticism signs.
All the signs for the 4th trait, neuroticism were flipped. It was previously measuring Emotional Stability. It might sound more positive, but it precise. Neuroticism is sensitivity to negative emotion, and that is a more precise trait to measure than stability because it is easier to recognize and define sensitivity to negative emotion rather than being insensitive to it.